perf: cast audio scaling constants to np.float32 to avoid float64 upcasting
Dividing a np.float32 array by a Python float triggers NumPy 2.0 implicit promotion to float64, doubling memory and degrading speed. Cast the constants explicitly to np.float32 to stay in the correct dtype throughout.

+++ src/nodetool/workflows/processing_context.py
@@ -1458,7 +1458,7 @@ class ProcessingContext:
 
         # Convert to target dtype if needed
         if dtype == "float32" and samples.dtype == np.int16:
-            samples = samples.astype(np.float32) / 32768.0
+            samples = samples.astype(np.float32) / np.float32(32768.0)
         elif dtype == "float64" and samples.dtype == np.int16:
             samples = samples.astype(np.float64) / 32768.0
         elif dtype == "int16" and samples.dtype in (np.float32, np.float64):
